LayoutTests/media/video-src-source.html

<video src=content/test.mp4 controls>
    <source src=content/error.mpeg>
</video>
<script src=video-test.js></script>
<script>
waitForEventTestAndEnd('begin', "relativeURL(video.currentSrc) == 'content/test.mp4'");
</script>
