LayoutTests/media/video-src-source-expected.txt

EVENT(begin) TEST(relativeURL(video.currentSrc) == 'content/test.mp4') FAIL
END OF TEST
